pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cbadvanced"
version = "3.0.0"
description = "The unofficial Python client for the Coinbase Advanced Trade API"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Magnus Lindland", email = "magnuslindland@proton.me" }
]
requires-python = ">=3.10"
keywords = [
    "coinbase",
    "coinbase-advanced",
    "trading",
    "cryptocurrency",
    "bitcoin",
    "ethereum",
    "api",
    "wrapper",
    "async",
    "httpx",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27",
    "PyJWT>=2.8",
    "cryptography>=42.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.urls]
Homepage = "https://github.com/mlindland/python-cbadvanced"
Documentation = "https://github.com/mlindland/python-cbadvanced#readme"
Repository = "https://github.com/mlindland/python-cbadvanced.git"
Issues = "https://github.com/mlindland/python-cbadvanced/issues"
Changelog = "https://github.com/mlindland/python-cbadvanced/blob/master/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["cbadvanced"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "SIM117", # nested with statements (clearer in tests)
]

[tool.ruff.lint.isort]
known-first-party = ["cbadvanced"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
